requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
albumentations==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
blessed==1.20.0
blinker==1.7.0
boto3==1.34.34
botocore==1.34.34
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
croniter==1.4.1
cycler==0.12.1
databricks-sdk==0.44.1
dateutils==0.6.12
deepdiff==6.7.1
Deprecated==1.2.18
docker==6.1.3
docstring-parser==0.15
editor==1.6.6
entrypoints==0.4
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.13.1
flask==3.0.2
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.60.1
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.26.2
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
inquirer==3.2.3
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonargparse==4.27.4
kiwisolver==1.4.5
kornia==0.7.4
kornia-rs==0.1.7
lazy-loader==0.3
lightning==2.2.0
lightning-api-access==0.0.5
lightning-cloud==0.5.64
lightning-fabric==2.2.0
lightning-utilities==0.10.1
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mlflow==2.20.3
mlflow-skinny==2.20.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.0
peft==0.13.2
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic-core==2.16.2
pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytorch-lightning==2.2.0
pytz==2023.4
PyYAML==6.0.1
qudida==0.0.4
querystring-parser==1.2.4
readchar==4.0.5
redis==5.0.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
runs==1.2.2
s3fs==2023.12.2
s3transfer==0.10.0
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.26
sqlparse==0.4.4
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2024.1.30
timm==0.9.12
tokenizers==0.20.3
torch==2.2.0
torchmetrics==1.3.0.post0
torchvision==0.17.0
tqdm==4.66.2
traitlets==5.14.1
transformers==4.46.2
triton==2.2.0
types-python-dateutil==2.8.19.20240106
typeshed-client==2.4.0
typing-extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.27.1
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
werkzeug==3.0.1
wrapt==1.16.0
xmod==1.8.1
yarl==1.9.4
zipp==3.17.0
